[Web/Endpoints] Moved Demo related API to their own controller
